+++ src/components/ModalContainer/type.ts
@@ -1,0 +1,4 @@
+export interface ModalContainerProps {
+  children: React.ReactNode;
+  closeModal: () => void;
+}
