+++ src/components/RestaurantDetailModal/index.css
@@ -1,29 +1,3 @@
-.modal {
-  display: block;
-}
-
-.modal-backdrop {
-  position: fixed;
-  top: 0;
-  right: 0;
-  bottom: 0;
-  left: 0;
-
-  background: rgba(0, 0, 0, 0.35);
-}
-
-.modal-container {
-  position: fixed;
-  bottom: 0;
-  max-height: 90%;
-  width: 100%;
-
-  padding: 32px 16px;
-
-  border-radius: 8px 8px 0px 0px;
-  background: var(--grey-100);
-}
-
 .restaurant-detail-container {
   display: flex;
   flex-direction: column;
